python_backend/requirements.txt
uvicorn
fastapi
pydantic
google-generativeai
python-dotenv
